# core: allocate less garbage iterators in Metadata.serialize