@carl-mastrangelo 2016-05-25
This was found from the memory analysis of https://github.com/grpc/grpc-java/pull/1856

+cc: @buchgr